handled order creation event
